+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix(tui): change that the default `add_random_(song|album)` keys were inverted.
 - Fix(tui): change Theme preview to not reset to index 0 each preview.
 - Fix(tui): not having the current theme selected when entering Theme preview tab.
+- Fix(tui): actually report any errors when adding to the playlist. (Like "invalid file type")
 
 ### [V0.9.1]
 - Released on: August 21, 2024.